Feat: utils
