+++ includes/utils.h
@@ -1,0 +1,8 @@
+#ifndef __UTILS_H__
+#define __UTILS_H__
+
+char		*gettodaydate();
+long long	getfocustime(char *user, char *date);
+char*		strjoin(const char* str1, const char* str2);
+
+#endif
